+++ pcs/test/test_cluster_pcmk_remote.py
@@ -300,6 +300,9 @@ class NodeAddGuest(ResourceTest):
         )
 
     def test_success_when_guest_node_matches_with_existing_guest(self):
+        # This test belongs to pcs/test/test_resource.py as it tests
+        # "resource update". But due to some fixtures it is more practical to
+        # keep it here.
         self.create_resource()
         self.assert_pcs_success(
             "cluster node add-guest node-host G",