+++ pcsd/pcs.rb
@@ -1594,7 +1594,10 @@ def cluster_status_from_nodes(auth_user, cluster_nodes, cluster_name)
     :status => 'unknown',
     :node_list => [],
     :resource_list => [],
+    # deprecated, kept for backward compatibility
+    # use pcsd_capabilities instead
     :available_features => [],
+    :pcsd_capabilities => [],
   }
 
   threads = []
@@ -1624,6 +1627,7 @@ def cluster_status_from_nodes(auth_user, cluster_nodes, cluster_name)
       begin
         parsed_response = JSON.parse(response, {:symbolize_names => true})
         parsed_response[:available_features] ||= []
+        parsed_response[:pcsd_capabilities] ||= []
         if parsed_response[:noresponse]
           node_map[node][:node] = {}
           node_map[node][:node].update(node_status_unknown)
@@ -1714,10 +1718,14 @@ def cluster_status_from_nodes(auth_user, cluster_nodes, cluster_name)
   node_map.each { |_, cluster_status|
     node_status = cluster_status[:node][:status]
     node_name = cluster_status[:node][:name]
-    # create set of available features on all nodes
-    # it is intersection of available features from all nodes
+    # Create a set of available features on all nodes as an intersection of
+    # available features from all nodes. Do it for both the old deprecated list
+    # (available_features) and the new one (pcsd_capabilities)
     if node_status != 'unknown' and cluster_status[:available_features]
       status[:available_features] &= cluster_status[:available_features]
+    end
+    if node_status != 'unknown' and cluster_status[:pcsd_capabilities]
+      status[:pcsd_capabilities] &= cluster_status[:pcsd_capabilities]
     end
     if (
       cluster_status[:node][:services] and
@@ -1861,6 +1869,8 @@ def get_node_status(auth_user, cib_dom)
       :nodes_utilization => get_nodes_utilization(cib_dom),
       :alerts => get_alerts(auth_user),
       :known_nodes => [],
+      # deprecated, kept for backward compatibility
+      # use pcsd_capabilities instead
       :available_features => [
         'constraint_colocation_set',
         'sbd',
@@ -1869,7 +1879,8 @@ def get_node_status(auth_user, cib_dom)
         'unmanaged_resource',
         'alerts',
         'hardened_cluster',
-      ]
+      ],
+      :pcsd_capabilities => CAPABILITIES_PCSD
   }
 
   nodes = get_nodes_status()